apps/ai/requirements.txt
langchain==0.2.1
langchain-openai==0.1.8
faiss-cpu==1.8.0
python-dotenv==1.0.1
pydantic==2.7.4
fastapi==0.111.0
uvicorn==0.30.1
